Fixed problem with small integer on tag

+++ src/client.h
@@ -18,7 +18,7 @@ class Client {
         ~Client();
         void send(std::string message);
         void connect();
-        std::pair<std::string,bool> receive(int tag);
+        std::pair<std::string,bool> receive(uint64_t tag);
         std::string simple_receive(int tag, std::string sent_command);
         void init_openssl(std::string cert_file_, std::string cert_dir_);
         bool verify_certificate();
